refactor: move BulkChangePreviewAPI from rest to housekeeping
