+++ housekeeping/src/main/groovy/whelk/HouseKeepingServer.java
@@ -3,6 +3,7 @@ package whelk;
 import org.eclipse.jetty.ee8.servlet.ServletContextHandler;
 import org.eclipse.jetty.ee8.servlet.ServletHolder;
 import org.eclipse.jetty.server.Server;
+import whelk.housekeeping.BulkChangePreviewAPI;
 import whelk.housekeeping.WebInterface;
 
 public class HouseKeepingServer extends XlServer {
@@ -16,6 +17,7 @@ public class HouseKeepingServer extends XlServer {
         ServletHolder holder = new ServletHolder(WebInterface.class);
         holder.setInitOrder(0);
         context.addServlet(holder, "/");
+        context.addServlet(BulkChangePreviewAPI.class, "/_bulk-change/*");
 
         serveStaticContent(context);
     }
