+++ rest/src/main/java/whelk/RestServer.java
@@ -10,7 +10,6 @@ import org.eclipse.jetty.http.UriCompliance;
 import org.eclipse.jetty.rewrite.handler.RewriteHandler;
 import org.eclipse.jetty.server.HttpConnectionFactory;
 import org.eclipse.jetty.server.Server;
-import whelk.rest.api.BulkChangePreviewAPI;
 import whelk.rest.api.ConverterAPI;
 import whelk.rest.api.Crud;
 import whelk.rest.api.DuplicatesAPI;
@@ -101,8 +100,6 @@ public class RestServer extends XlServer {
         context.addServlet(HoldAPI.class, "/_findhold");
         context.addServlet(RecordRelationAPI.class, "/_dependencies");
         context.addServlet(DuplicatesAPI.class, "/_duplicates");
-
-        context.addServlet(BulkChangePreviewAPI.class, "/_bulk-change/*");
 
         context.addServlet(se.kb.libris.digi.DigitalReproductionAPI.class, "/_reproduction");
